FEATURE: Alternative interpreters (#36)
* In pull request #33 I wrote the line handling the printing of stderr in
the script run function so in the current state it doesn't print stderr.
This is a patch to fix that, it also has a regression test included to
future proof it. Also fixed the verbose ending line. I was working on
this during the night so it seems I was a bit tired and missed a few
things.

* Feature: alternative interpreters

* If shebang found on the first line create a tempoarary executable file. otherwise use the default interpreter.
* Added tests for scripts with shebang and the default_interpreter
  config option.
* Updated README to document the ability to use alternative
  interpreters.